requirements.txt
# Core dependencies
yt-dlp>=2024.0.0
Pillow>=10.0.0
imagehash>=4.3.0

# Fast transcription (4x faster than openai-whisper)
faster-whisper>=1.0.0

# Development
pytest>=8.0.0
build>=1.0.0
